Refactor: Split CustomChat component
Split the CustomChat component into smaller, more manageable components to improve readability and maintainability.
