+++ src/components/CustomChat.tsx
@@ -1,19 +1,14 @@
 
 import { useState, useEffect, useRef } from 'react';
 import { v4 as uuidv4 } from 'uuid';
-import { Button } from './ui/button';
-import { Input } from './ui/input';
+import { MessageCircle } from 'lucide-react';
 import { Separator } from './ui/separator';
-import { MessageCircle, Send, Calendar, Mic } from 'lucide-react';
 import { useDialog } from './providers/DialogProvider';
-
-// Message type definition
-interface Message {
-  id: string;
-  text: string;
-  sender: 'user' | 'bot';
-  timestamp: Date;
-}
+import { Message } from './chat/ChatMessage';
+import ChatHeader from './chat/ChatHeader';
+import ChatMessages from './chat/ChatMessages';
+import ChatInput from './chat/ChatInput';
+import { sendChatMessage, addSystemMessage } from '../services/chatService';
 
 const CustomChat = () => {
   const [isOpen, setIsOpen] = useState(false);
@@ -21,7 +16,6 @@ const CustomChat = () => {
   const [inputValue, setInputValue] = useState('');
   const [sessionId, setSessionId] = useState('');
   const [isLoading, setIsLoading] = useState(false);
-  const messagesEndRef = useRef<HTMLDivElement>(null);
   const chatContainerRef = useRef<HTMLDivElement>(null);
   const { setCalendarDialogOpen } = useDialog();
 
@@ -40,12 +34,7 @@ const CustomChat = () => {
     // Add initial bot messages if this is a new session
     if (!storedSessionId) {
       const initialMessages: Message[] = [
-        {
-          id: uuidv4(),
-          text: "👋 Hallo! Ich bin dein Assistent.",
-          sender: 'bot',
-          timestamp: new Date()
-        },
+        addSystemMessage("👋 Hallo! Ich bin dein Assistent."),
         {
           id: uuidv4(),
           text: "Du kannst direkt einen Termin buchen oder mir eine Frage stellen!",
@@ -58,85 +47,23 @@ const CustomChat = () => {
     }
   }, []);
 
-  // Scroll to bottom when messages change
-  useEffect(() => {
-    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
-  }, [messages]);
-
-  // Handle sending messages to n8n
-  const sendMessage = async (text: string) => {
-    if (!text.trim()) return;
-    
-    // Add user message to chat
-    const userMessage: Message = {
-      id: uuidv4(),
-      text,
-      sender: 'user',
-      timestamp: new Date()
-    };
-    
-    setMessages(prev => [...prev, userMessage]);
+  // Handle sending messages
+  const handleSendMessage = async (text: string) => {
+    await sendChatMessage(text, sessionId, setMessages, setIsLoading);
     setInputValue('');
-    setIsLoading(true);
-    
-    try {
-      const response = await fetch('https://automatisierung.seserver.nohype-ai.de/webhook/f406671e-c954-4691-b39a-66c90aa2f103/chat', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'Accept': 'application/json'
-        },
-        body: JSON.stringify({
-          chatInput: text,
-          sessionId: sessionId,
-          metadata: {
-            source: 'website',
-            page: window.location.pathname
-          }
-        })
-      });
-      
-      const data = await response.json();
-      
-      if (data.output) {
-        // Add bot response to chat
-        const botMessage: Message = {
-          id: uuidv4(),
-          text: data.output,
-          sender: 'bot',
-          timestamp: new Date()
-        };
-        
-        setMessages(prev => [...prev, botMessage]);
-      }
-    } catch (error) {
-      console.error('Error sending message:', error);
-      
-      // Add error message
-      const errorMessage: Message = {
-        id: uuidv4(),
-        text: "Entschuldigung, es gab ein Problem bei der Verbindung. Bitte versuche es später erneut.",
-        sender: 'bot',
-        timestamp: new Date()
-      };
-      
-      setMessages(prev => [...prev, errorMessage]);
-    } finally {
-      setIsLoading(false);
-    }
   };
 
   // Handle input submission
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault();
-    sendMessage(inputValue);
+    handleSendMessage(inputValue);
   };
 
   // Handle keyboard shortcut (Enter to send)
   const handleKeyDown = (e: React.KeyboardEvent) => {
     if (e.key === 'Enter' && !e.shiftKey) {
       e.preventDefault();
-      sendMessage(inputValue);
+      handleSendMessage(inputValue);
     }
   };
 
@@ -149,36 +76,9 @@ const CustomChat = () => {
   const handleVoiceChat = () => {
     console.log("Sprachchat");
     // Placeholder for voice chat functionality
-    const voiceMessage: Message = {
-      id: uuidv4(),
-      text: "Sprachchat ist aktuell noch in Entwicklung. Bitte versuche es später erneut.",
-      sender: 'bot',
-      timestamp: new Date()
-    };
+    const voiceMessage = addSystemMessage("Sprachchat ist aktuell noch in Entwicklung. Bitte versuche es später erneut.");
     
     setMessages(prev => [...prev, voiceMessage]);
-  };
-
-  // Render a message bubble
-  const renderMessage = (message: Message) => {
-    const isBot = message.sender === 'bot';
-    
-    return (
-      <div 
-        key={message.id} 
-        className={`flex ${isBot ? 'justify-start' : 'justify-end'} mb-4`}
-      >
-        <div 
-          className={`rounded-2xl px-4 py-2 max-w-[80%] shadow-md ${
-            isBot 
-              ? 'bg-secondary/10 text-gray-100' 
-              : 'bg-primary/90 text-white'
-          }`}
-        >
-          {message.text}
-        </div>
-      </div>
-    );
   };
 
   // Toggle chat window
@@ -207,77 +107,27 @@ const CustomChat = () => {
           style={{ backdropFilter: 'blur(10px)' }}
         >
           {/* Chat header */}
-          <div className="flex items-center justify-between p-4 bg-gradient-to-r from-secondary/80 to-primary/80">
-            <div className="flex items-center space-x-2">
-              <div className="text-white font-bold">AI Assistent</div>
-            </div>
-            <div className="flex flex-col space-y-2">
-              <Button 
-                onClick={handleBookAppointment}
-                size="sm" 
-                variant="outline"
-                className="text-xs bg-black/20 text-white border-white/20 hover:bg-black/40"
-              >
-                <Calendar className="mr-1 h-3 w-3" /> Termin buchen
-              </Button>
-              <Button 
-                onClick={handleVoiceChat}
-                size="sm" 
-                variant="outline"
-                className="text-xs bg-black/20 text-white border-white/20 hover:bg-black/40"
-              >
-                <Mic className="mr-1 h-3 w-3" /> Sprachchat
-              </Button>
-              <Button 
-                onClick={toggleChat} 
-                variant="ghost"
-                size="sm"
-                className="text-white hover:bg-black/20 p-1"
-              >
-                ✕
-              </Button>
-            </div>
-          </div>
+          <ChatHeader 
+            handleBookAppointment={handleBookAppointment}
+            handleVoiceChat={handleVoiceChat}
+            toggleChat={toggleChat}
+          />
           
           <Separator />
           
           {/* Messages container */}
-          <div className="flex-1 p-4 overflow-y-auto bg-black/60">
-            {messages.map(renderMessage)}
-            {isLoading && (
-              <div className="flex justify-start mb-4">
-                <div className="bg-secondary/10 text-gray-100 rounded-2xl px-4 py-2 max-w-[80%]">
-                  <div className="flex space-x-1">
-                    <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0s' }}></div>
-                    <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0.2s' }}></div>
-                    <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0.4s' }}></div>
-                  </div>
-                </div>
-              </div>
-            )}
-            <div ref={messagesEndRef} />
-          </div>
+          <ChatMessages messages={messages} isLoading={isLoading} />
           
           <Separator />
           
           {/* Chat input */}
-          <form onSubmit={handleSubmit} className="p-3 flex">
-            <Input
-              value={inputValue}
-              onChange={(e) => setInputValue(e.target.value)}
-              onKeyDown={handleKeyDown}
-              placeholder="Was möchtest du wissen?"
-              className="flex-1 bg-black/40 border-gray-700 text-white focus-visible:ring-primary"
-              disabled={isLoading}
-            />
-            <Button 
-              type="submit" 
-              className="ml-2" 
-              disabled={isLoading || !inputValue.trim()}
-            >
-              <Send className="h-4 w-4" />
-            </Button>
-          </form>
+          <ChatInput 
+            inputValue={inputValue}
+            setInputValue={setInputValue}
+            handleSubmit={handleSubmit}
+            handleKeyDown={handleKeyDown}
+            isLoading={isLoading}
+          />
         </div>
       )}
     </>
