+++ src/components/chat/ChatMessages.tsx
@@ -1,0 +1,39 @@
+
+import { useRef, useEffect } from 'react';
+import ChatMessage, { Message } from './ChatMessage';
+
+interface ChatMessagesProps {
+  messages: Message[];
+  isLoading: boolean;
+}
+
+const ChatMessages = ({ messages, isLoading }: ChatMessagesProps) => {
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+
+  // Scroll to bottom when messages change
+  useEffect(() => {
+    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+  }, [messages]);
+
+  return (
+    <div className="flex-1 p-4 overflow-y-auto bg-black/60">
+      {messages.map((message) => (
+        <ChatMessage key={message.id} message={message} />
+      ))}
+      {isLoading && (
+        <div className="flex justify-start mb-4">
+          <div className="bg-secondary/10 text-gray-100 rounded-2xl px-4 py-2 max-w-[80%]">
+            <div className="flex space-x-1">
+              <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0s' }}></div>
+              <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0.2s' }}></div>
+              <div className="w-2 h-2 bg-gray-400 rounded-full animate-bounce" style={{ animationDelay: '0.4s' }}></div>
+            </div>
+          </div>
+        </div>
+      )}
+      <div ref={messagesEndRef} />
+    </div>
+  );
+};
+
+export default ChatMessages;
